refactor: rewrite cjsToEsm.mjs  in ECMAScript
